Added docker compose file with useful tweaks. (#385)
Summary:
From my perspective a docker-compose file has several benefits. On the one hand, it increases the comfort and on the other hand it is a way to supply users with useful tweaks. The commited docker-compose file addresses several issues and tweaks:
1. It fixes potential problems with dataloaders (see https://github.com/facebookresearch/detectron2/issues/384).
2. It includes Multi-GPU and performance tweaks as suggested by NVIDIA (see https://docs.nvidia.com/deeplearning/frameworks/user-guide/index.html#caffeovr).
3. It adds GUI support (see https://github.com/facebookresearch/detectron2/issues/379).
4. It enables caching of downloaded models (see https://github.com/facebookresearch/detectron2/issues/382).
5. It makes docker run with the UID of the host user.

Of course, all this can be accomplished with a long docker command. However, a docker-compose file give a central place to gather all recommendations for running detectron2 with docker, without bloating the Dockerfile with comments.
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/385

Differential Revision: D18867052

Pulled By: ppwwyyxx

fbshipit-source-id: 2576541545ef73043888c6e7201f4452d7d11d31